+++ Source/Geom/Geometry.cpp
@@ -18,7 +18,7 @@ float GetXZAngle(const Vec3f& v1, const Vec3f& v2)
 
   float dx = v2.x - v1.x;
   float dz = v2.z - v1.z;
-  float r = atan2(dx, dz); // TODO do we need to check if z = 0 ?
+  float r = static_cast<float>(atan2(dx, dz)); // TODO do we need to check if z = 0 ?
   float rDegs = r / (float)M_PI * 180.0f; // convert to degs
   rDegs += 180.0f; // need this to fix z direction 
   return rDegs;
